# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GetIllustra Admin API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, published]
        - name: category_id
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectList"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    put:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Project deleted
  /projects/{id}/publish:
    post:
      operationId: publishProject
      tags: [projects]
      summary: Publish a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}/unpublish:
    post:
      operationId: unpublishProject
      tags: [projects]
      summary: Unpublish a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project unpublished
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}/toggle-featured:
    post:
      operationId: toggleProjectFeatured
      tags: [projects]
      summary: Toggle featured status of a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Featured toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  # ASSETS
...
  /styles:
    get:
      operationId: listStyles
      tags: [styles]
      summary: List all styles
      responses:
        "200":
          description: List of styles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Style"
    post:
      operationId: createStyle
      tags: [styles]
      summary: Create a style
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStyleBody"
      responses:
        "201":
          description: Style created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Style"
  /styles/{id}:
    put:
      operationId: updateStyle
      tags: [styles]
      summary: Update a style
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStyleBody"
      responses:
        "200":
          description: Updated style
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Style"
    delete:
      operationId: deleteStyle
      tags: [styles]
      summary: Delete a style
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Style deleted

  # TAGS
...
components:
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        coverImage:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, published]
        featured:
          type: boolean
        categoryId:
          type: string
          nullable: true
        categoryName:
          type: string
          nullable: true
        styleIds:
          type: array
          items:
            type: string
        tagIds:
          type: array
          items:
            type: string
        viewCount:
          type: integer
        downloadCount:
          type: integer
        assetCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - slug
        - status
        - featured
        - viewCount
        - downloadCount
        - assetCount
        - createdAt
        - updatedAt
        - styleIds
        - tagIds
    ProjectList:
      type: object
      properties:
        projects:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        total:
          type: integer
      required:
        - projects
        - total
    CreateProjectBody:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        coverImage:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, published]
        featured:
          type: boolean
        categoryId:
          type: string
          nullable: true
        styleIds:
          type: array
          items:
            type: string
        tagIds:
          type: array
          items:
            type: string
      required:
        - title
        - slug
        - status
        - featured
        - styleIds
        - tagIds
    UpdateProjectBody:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        coverImage:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, published]
        featured:
          type: boolean
        categoryId:
          type: string
          nullable: true
        styleIds:
          type: array
          items:
            type: string
        tagIds:
          type: array
          items:
            type: string
...
    Style:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        projectCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - slug
        - projectCount
        - createdAt
    CreateStyleBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
      required:
        - name
        - slug
    UpdateStyleBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
...