# lib/db/src/schema/styles.ts
import { pgTable, text, uuid, timestamp } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const stylesTable = pgTable("styles", {
  id: uuid("id").primaryKey().defaultRandom(),
  name: text("name").notNull(),
  slug: text("slug").notNull().unique(),
  createdAt: timestamp("created_at", { withTimezone: true }).notNull().defaultNow(),
  updatedAt: timestamp("updated_at", { withTimezone: true }).notNull().defaultNow().$onUpdate(() => new Date()),
});

export const insertStyleSchema = createInsertSchema(stylesTable).omit({ id: true, createdAt: true, updatedAt: true });
export type InsertStyle = z.infer<typeof insertStyleSchema>;
export type Style = typeof stylesTable.$inferSelect;